requirements.txt
bce-python-sdk==0.9.6
bcrypt==4.1.1
cryptography==42.0.5
fastapi==0.110.1
fastapi-cache2==0.2.1
fastjsonschema==2.19.1
passlib==1.7.4
pillow==10.3.0
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
python-jose==3.3.0
python-multipart==0.0.7
requests==2.32.2
SQLAlchemy==2.0.25
thefuzz==0.22.1
python-dateutil~=2.9.0.post0
paddleocr~=2.7.3
slowapi~=0.1.9
psycopg2-binary
pydantic[email]